CSV Parsing overhaul. Timetable no longer relies on a CSV parsing library to read GTFS archives. Instead, it manually parses entries using `getline` with a comma delimiter and special-cases quoted string entries.
The decision to manually parse files was made because GTFS is a loose standard, meaning it has many optional fields that change the parsing requirements (number of columns, column order, cell formatting).

Getting the manual parsing to work also implied some minor changes to `field_mapper`, but these are negligible, as their only use case is in CSV parsing.

+++ include/gtfs/shape.h
@@ -14,13 +14,13 @@ namespace gtfs {
       // Unique identifier for a shape
       std::string id;
       // Latitude of this point in the shape
-      float latitude;
+      double latitude;
       // Longitude of this point in the shape
-      float longitude;
+      double longitude;
       // Index of this point in the sequence of points for the shape
       int index;
       // Distance traveled from the previous shape point to get to this point
-      float distance;
+      double distance;
 
       shape() = default;
